python/requirements.txt
Flask
pylint
pytest
Pillow==9.5.0
moviepy
flask-cors
